+++ packages/nimble-components/src/rich-text/editor/styles.ts
@@ -1,5 +1,5 @@
 import { css } from '@microsoft/fast-element';
-import { display } from '@microsoft/fast-foundation';
+import { display } from '../../utilities/style/display';
 import {
     bodyDisabledFontColor,
     bodyFont,