+++ packages/nimble-components/src/select/testing/select.pageobject.ts
@@ -10,6 +10,7 @@ import type { Select } from '..';
 import type { ListOption } from '../../list-option';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { FilterMode } from '../types';
+import type { Button } from '../../button';
 
 /**
  * Page object for the `nimble-select` component to provide consistent ways
@@ -128,6 +129,26 @@ export class SelectPageObject {
         this.clickOption(optionIndex);
     }
 
+    public clickClearButton(): void {
+        if (!this.selectElement.clearable) {
+            throw new Error(
+                'Select must set "clearable" in order to click clear button'
+            );
+        }
+
+        if (
+            this.selectElement.selectedIndex === -1
+            || this.selectElement.displayPlaceholder
+        ) {
+            throw new Error(
+                'Select must have a selected element in order to click clear button'
+            );
+        }
+
+        const clearButton = this.getClearButton();
+        clearButton?.click();
+    }
+
     public async clickAway(): Promise<void> {
         this.selectElement.dispatchEvent(new Event('focusout'));
         await waitForUpdatesAsync();
@@ -218,6 +239,10 @@ export class SelectPageObject {
         );
     }
 
+    public isClearButtonVisible(): boolean {
+        return this.getClearButton() != null;
+    }
+
     public isFilterInputVisible(): boolean {
         return (
             this.selectElement.shadowRoot?.querySelector('.filter-field')
@@ -246,20 +271,6 @@ export class SelectPageObject {
         return this.selectElement.filterInput?.value ?? '';
     }
 
-    public async waitForChange(): Promise<string> {
-        return new Promise(resolve => {
-            this.selectElement.addEventListener(
-                'change',
-                () => {
-                    resolve(this.selectElement.value);
-                },
-                {
-                    once: true
-                }
-            );
-        });
-    }
-
     private getFilterInput(): HTMLInputElement | null | undefined {
         if (this.selectElement.filterMode === FilterMode.none) {
             throw new Error(
@@ -267,5 +278,11 @@ export class SelectPageObject {
             );
         }
         return this.selectElement.filterInput;
+    }
+
+    private getClearButton(): Button | null | undefined {
+        return this.selectElement.shadowRoot?.querySelector<Button>(
+            '.clear-button'
+        );
     }
 }